+++ engine/src/main/java/net/jqwik/engine/execution/reporting/MapValueReport.java
@@ -15,7 +15,7 @@ class MapValueReport extends ValueReport {
 	}
 
 	@Override
-	String singleLineReport() {
+	public String singleLineReport() {
 		return label.orElse("") + "{" + singleLineEntries() + "}";
 	}
 
@@ -28,7 +28,7 @@ class MapValueReport extends ValueReport {
 	}
 
 	@Override
-	void report(LineReporter lineReporter, int indentLevel, String appendix) {
+	public void report(LineReporter lineReporter, int indentLevel, String appendix) {
 		lineReporter.addLine(indentLevel, label.orElse("") + "{");
 		reportEntries(lineReporter, indentLevel + 1);
 		lineReporter.addLine(indentLevel, "}" + appendix);